+++ src/main/kotlin/com/github/shmvanhouten/adventofcode2019/day22/ShuffleInstruction.kt
@@ -1,6 +1,6 @@
 package com.github.shmvanhouten.adventofcode2019.day22
 
-data class ShuffleInstruction(val type: Instructiontype, val number: Int = -1)
+data class ShuffleInstruction(val type: Instructiontype, val number: Long = -1L)
 
 enum class Instructiontype {
     DEAL_WITH_INCREMENT,
